requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
Django==3.0.6
djangorestframework==3.11.0
flake8==3.8.1
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pluggy==0.13.1
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylint==2.5.2
python-jsonrpc-server==0.3.4
python-language-server==0.32.0
pytz==2020.1
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
ujson==1.35
wrapt==1.12.1
yapf==0.30.0
